:sparkles: fetch new withings token once every 3 hours
should fix #42
